[ghstack-poisoned]

+++ comms/torchcomms/BackendWrapper.cpp
@@ -8,6 +8,10 @@ namespace comms {
 
 namespace {
 
+// Extract the scaling factor from NCCL's PREMUL_SUM operation supplement.
+// NCCLPreMulSumSupplement stores either a tensor or double scaling factor
+// that is applied before summation. The reinterpret_cast is safe because
+// PREMUL_SUM operations always have a NCCLPreMulSumSupplement attached.
 PreMulSumFactorT getPreMulSumFactor(const c10d::ReduceOp& op) {
   const auto* preMulSupplement =
       reinterpret_cast<c10d::NCCLPreMulSumSupplement*>(op.supplement_.get());
@@ -58,10 +62,17 @@ bool WorkWrapper::isCompleted() {
   return work_->isCompleted();
 }
 bool WorkWrapper::isSuccess() const {
-  // TODO: implement error states
+  // Note: Error state tracking is not implemented. This method returns
+  // isCompleted() as a simplification. The underlying TorchWork does not
+  // expose separate success/error states, so we assume completion implies
+  // success. Callers that need error detection should use try/catch around
+  // wait() instead.
   return work_->isCompleted();
 }
 std::exception_ptr WorkWrapper::exception() const {
+  // Note: Exception capture is not implemented. The underlying TorchWork
+  // interface does not provide a mechanism to retrieve exceptions after
+  // completion. Errors are raised during wait() calls instead.
   return nullptr;
 }
 bool WorkWrapper::wait(std::chrono::milliseconds timeout) {
@@ -72,7 +83,10 @@ bool WorkWrapper::wait(std::chrono::milliseconds timeout) {
   return true;
 }
 void WorkWrapper::synchronize() {
-  // TODO: this should only wait on stream
+  // Note: Ideally this should only synchronize on the CUDA stream without
+  // blocking the CPU thread. However, the current TorchWork interface does
+  // not expose stream-only synchronization, so we fall back to a full wait()
+  // which blocks until the operation completes.
   return work_->wait();
 }
 std::vector<at::Tensor> WorkWrapper::result() {
@@ -93,8 +107,8 @@ c10::intrusive_ptr<c10d::Work> BackendWrapper::broadcast(
   if (opts.timeout != kUnsetTimeout) {
     bopts.timeout = opts.timeout;
   }
-  return c10::make_intrusive<WorkWrapper>(
-      backend_->broadcast(tensors.at(0), opts.rootRank, opts.asyncOp, bopts));
+  return c10::make_intrusive<WorkWrapper>(backend_->broadcast(
+      tensors.at(0), static_cast<int>(opts.rootRank), opts.asyncOp, bopts));
 }
 
 c10::intrusive_ptr<c10d::Work> BackendWrapper::allreduce(
@@ -131,7 +145,7 @@ c10::intrusive_ptr<c10d::Work> BackendWrapper::reduce(
   }
   return c10::make_intrusive<WorkWrapper>(backend_->reduce(
       tensors.at(0),
-      opts.rootRank,
+      static_cast<int>(opts.rootRank),
       toReduceOp(opts.reduceOp),
       opts.asyncOp,
       bopts));
@@ -169,9 +183,9 @@ c10::intrusive_ptr<c10d::Work> BackendWrapper::allgather_coalesced(
       outputTensorLists.at(0), inputTensors.at(0), opts.asyncOp, bopts));
 }
 
-// TODO: Need to implement the case when input/output tensors are larger than
-// one. since this a coalesced version. We only support one input/output tensor
-// for now.
+// Note: Coalesced operations with multiple input/output tensors are not yet
+// supported. Currently only single tensor is supported. When extending this,
+// iterate over all tensors and coalesce them into a single backend call.
 c10::intrusive_ptr<c10d::Work> BackendWrapper::allgather_into_tensor_coalesced(
     std::vector<at::Tensor>& output_tensors,
     std::vector<at::Tensor>& inputTensors,
@@ -213,7 +227,10 @@ c10::intrusive_ptr<c10d::Work> BackendWrapper::gather(
     bopts.timeout = opts.timeout;
   }
   return c10::make_intrusive<WorkWrapper>(backend_->gather(
-      outputTensors.at(0), inputTensors.at(0), opts.rootRank, opts.asyncOp));
+      outputTensors.at(0),
+      inputTensors.at(0),
+      static_cast<int>(opts.rootRank),
+      opts.asyncOp));
 }
 
 c10::intrusive_ptr<c10d::Work> BackendWrapper::scatter(
@@ -234,7 +251,10 @@ c10::intrusive_ptr<c10d::Work> BackendWrapper::scatter(
     inputTensors.emplace_back();
   }
   return c10::make_intrusive<WorkWrapper>(backend_->scatter(
-      outputTensors.at(0), inputTensors.at(0), opts.rootRank, opts.asyncOp));
+      outputTensors.at(0),
+      inputTensors.at(0),
+      static_cast<int>(opts.rootRank),
+      opts.asyncOp));
 }
 
 c10::intrusive_ptr<c10d::Work> BackendWrapper::reduce_scatter(
@@ -257,9 +277,9 @@ c10::intrusive_ptr<c10d::Work> BackendWrapper::reduce_scatter(
       bopts));
 }
 
-// TODO: Need to implement the case when input/output tensors are larger than
-// one. since this a coalesced version. We only support one input/output tensor
-// for now.
+// Note: Coalesced operations with multiple input/output tensors are not yet
+// supported. Currently only single tensor is supported. When extending this,
+// iterate over all tensors and coalesce them into a single backend call.
 c10::intrusive_ptr<c10d::Work> BackendWrapper::reduce_scatter_tensor_coalesced(
     std::vector<at::Tensor>& outputTensors,
     std::vector<at::Tensor>& inputTensors,
